time class, fps counter

+++ src/engine/renderer/TextRenderer.cpp
@@ -150,7 +150,7 @@ namespace cologne
                 texture,
                 glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                 glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                face->glyph->advance.x
+                static_cast<uint32_t>(face->glyph->advance.x)
             };
             characters.insert(std::pair<char, Character>(c, character));
         }
